requirements.txt
docx2python==3.0.0
lxml==5.2.2
paragraphs==0.2.1
pillow==10.4.0
python-docx==1.1.2
typing-extensions==4.12.2
